refactor: core.py -> search.py + probes.py + analysis.py (mixins); core.py is a back-compat shim
All gates green: ruff, 16 unit, golden CdSe barrier in band.
