+++ pallas/__init__.py
@@ -1,9 +1,9 @@
 """PALLAS — Phase Transition Landscape Exploration with Automated Saddle Search."""
 
-from pallas.core import Pallas, PallasAtom, PallasConfig
 from pallas.graph import minimax_barrier, minimax_path, minimax_path_kinetic
 from pallas.nequip_calc import NequIPDualCalc
 from pallas.optimize import cal_saddle, local_optimization, set_calculator
+from pallas.search import Pallas, PallasAtom, PallasConfig
 from pallas.xcal import XCalculator, atoms_to_cell, fp_dist_with_assignment
 
 __all__ = [
